Move common logic to parent class
