+++ src/org/ballerinalang/plugins/idea/run/configuration/application/BallerinaApplicationRunConfiguration.java
@@ -21,16 +21,12 @@ import com.intellij.execution.Executor;
 import com.intellij.execution.configurations.ConfigurationFactory;
 import com.intellij.execution.configurations.RunConfiguration;
 import com.intellij.execution.configurations.RunProfileState;
-import com.intellij.execution.configurations.RuntimeConfigurationError;
-import com.intellij.execution.configurations.RuntimeConfigurationException;
 import com.intellij.execution.runners.ExecutionEnvironment;
 import com.intellij.openapi.options.SettingsEditor;
 import com.intellij.openapi.project.Project;
 import org.ballerinalang.plugins.idea.run.configuration.BallerinaApplicationRunningState;
 import org.ballerinalang.plugins.idea.run.configuration.BallerinaRunConfigurationBase;
-import org.ballerinalang.plugins.idea.run.configuration.BallerinaRunUtil;
 import org.ballerinalang.plugins.idea.run.configuration.ui.BallerinaApplicationSettingsEditor;
-import org.ballerinalang.plugins.idea.sdk.BallerinaSdkUtil;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -44,21 +40,6 @@ public class BallerinaApplicationRunConfiguration extends BallerinaRunConfigurat
     @Override
     public SettingsEditor<? extends RunConfiguration> getConfigurationEditor() {
         return new BallerinaApplicationSettingsEditor();
-    }
-
-    @Override
-    public void checkConfiguration() throws RuntimeConfigurationException {
-        String ballerinaExecutablePath = BallerinaSdkUtil.getBallerinaExecutablePath(getProject());
-        if (ballerinaExecutablePath.isEmpty()) {
-            throw new RuntimeConfigurationError("Cannot find Ballerina executable. Please check Ballerina SDK path.");
-        }
-        String openedFile = BallerinaRunUtil.getOpenFilePath(getProject());
-        if (openedFile.isEmpty()) {
-            throw new RuntimeConfigurationError("No Ballerina file is opened in the editor.");
-        }
-        if (!BallerinaRunUtil.isBallerinaFileOpen(getProject())) {
-            throw new RuntimeConfigurationError("Opened file in the editor is not a Ballerina file.");
-        }
     }
 
     @Nullable
